[FIX] mail: crash if no attachment
When the mail composer is opened without a default template, a traceback
is raised.

This is because `values['attachment_ids']` can either be a list of IDS
(if we go through the `if` and `elif` conditions) or already a command
(if we go through the `else` condition). In case of a command, we end up
with something like `[(6, 0, [(6, 0, )])]`, which causes issues.

opw-2003197
opw-2003130

closes odoo/odoo#33707

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>